prepare for new list view
